CEPBN1/requirements.txt
streamlit
opencv-python-headless
pandas
Pillow
numpy